Fix to Custom Radius mode of Hyperbolic Layout Object
I found a bug in the custom radius mode that was recently added to the hyperbolic layout object. Basically, the oth radius and properties of the inner ring are set at creation time, before the cost radii get specified and used, so if you overwrite the oth radius, then the changes don't propagate to all the new places.      I couldn't find an elegant fix, so I added a warning instead. Custom radius mode should only be used to set the outer radii before populating the graph. The size of the innermost ring is set at creation time of the object using the "side" parameter.